Fixed some VC++ 8.0 compiler warnings. (#2930)
git-svn-id: https://svn.osgeo.org/gdal/trunk@17405 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/aigrid/gridlib.c
@@ -910,7 +910,7 @@ CPLErr AIGReadBlockIndex( AIGInfo_t * psInfo, AIGTileInfo *psTInfo,
         return CE_Failure;
     }
     VSIFSeekL( fp, 100, SEEK_SET );
-    if (VSIFReadL( panIndex, 8, psTInfo->nBlocks, fp ) != psTInfo->nBlocks)
+    if ((int)VSIFReadL( panIndex, 8, psTInfo->nBlocks, fp ) != psTInfo->nBlocks)
     {
         CPLError(CE_Failure, CPLE_AppDefined,
                  "AIGReadBlockIndex: Cannot read block info");
@@ -1065,8 +1065,4 @@ CPLErr AIGReadStatistics( const char * pszCoverName, AIGInfo_t * psInfo )
 
     return( CE_None );
 }
-
-
-
-
